user_form_app/requirements.txt
# 9/8 Good to commit #
Flask==2.3.3
psycopg2-binary==2.9.7